requirements.txt
    altair                      # Declarative statistical visualization library for Python
    asyncio                     # Asynchronous I/O, event loop, and coroutine support
    beautifulsoup4              # Library for parsing HTML and XML documents
    bs4                         # Another name for BeautifulSoup, used for web scraping
    extra-streamlit-components  # Custom and extra components for Streamlit applications
    fastapi                     # Fast and modern web framework for building APIs
    ffmpeg                      # Tool for handling multimedia data
    folium                      # Python library for interactive maps using Leaflet.js
    google-api-python-client    # Google APIs Python client library
    google-search-results       # API for retrieving search results from Google
    googlemaps                  # Python client for Google Maps services
    IPython                     # Interactive computing environment for Python
    moviepy                     # Video editing library for Python
    nest-asyncio                # Allows nested use of asyncio.run and event loops
    numpy                       # Fundamental package for numerical computing
    opencv-python-headless      # OpenCV library for computer vision tasks without GUI
    pandas                      # Data manipulation and analysis library
    pathlib                     # Object-oriented filesystem paths
    pillow                      # Python Imaging Library (PIL) fork for image processing
    playwright                  # Browser automation library for Python
    plotly                      # Interactive graphing library
    plotly.express              # Simple syntax for complex Plotly visualizations
    pydantic                    # Data validation and settings management using Python type annotations
    pydub                       # Manipulate audio with an intuitive interface
    pygwalker                   # Visualization tool for exploring pandas DataFrames
    PyScrappy                   # Web scraping library for Python
    python-docx                 # python-docx is a Python library for creating and updating Microsoft Word (.docx) files.
    pytube                      # Python library for downloading YouTube videos
    pyzillow                    # Python wrapper for Zillow API
    requests                    # Simple HTTP library for Python
    scrapegraphai               # Scrape data from graphs and charts
    Scrapy                      # Web crawling and scraping framework
    simple-salesforce           # Simple Salesforce API wrapper for Python
    SpeechRecognition           # Library for performing speech recognition with Python
    streamlit                   # Framework for creating interactive web apps from Python scripts
    streamlit-antd-components   # Ant Design components for Streamlit apps
    streamlit-echarts           # ECharts components for Streamlit apps
    streamlit-elements          # Streamlit component library for building complex UIs
    streamlit-extras            # Additional components and utilities for Streamlit
    streamlit-folium            # Folium integration for Streamlit apps
    stripe                      # Python library for Stripe API integration
    supabase                    # Python client for Supabase, a Firebase alternative
    tavily-python               # Tavily Python client for machine learning tasks
    temp                        # Temporary file and directory management
    tiktoken                    # Tokenizer for OpenAI models
    typing                      # Type hints for Python
    urllib3                     # HTTP library with thread-safe connection pooling
    webscraping                 # General web scraping library
    WebScraperAPI               # API for web scraping tasks
    yelpapi                     # Python client for Yelp Fusion API
    youtube-transcript-api      # Retrieve YouTube video transcripts with Python
    youtube_dl                  # Download videos from YouTube and other sites
